# Conflicts:
#	src/main/kotlin/com/lykke/matching/engine/config/spring/DatabaseAccessorConfig.kt
#	src/main/kotlin/com/lykke/matching/engine/holders/BalancesHolder.kt
#	src/main/kotlin/com/lykke/matching/engine/messages/MessageProcessor.kt
#	src/main/kotlin/com/lykke/matching/engine/order/GenericLimitOrderProcessor.kt
#	src/main/kotlin/com/lykke/matching/engine/order/SingleLimitOrderProcessor.kt
#	src/main/kotlin/com/lykke/matching/engine/order/StopLimitOrderProcessor.kt
#	src/main/kotlin/com/lykke/matching/engine/services/BalanceUpdateService.kt
#	src/main/kotlin/com/lykke/matching/engine/services/CashTransferOperationService.kt
#	src/main/kotlin/com/lykke/matching/engine/services/ReservedBalanceUpdateService.kt
#	src/main/resources/application.properties
#	src/test/kotlin/com/lykke/matching/engine/balance/util/TestBalanceHolderWrapper.kt
#	src/test/kotlin/com/lykke/matching/engine/services/MarketOrderServiceTest.kt

+++ src/main/kotlin/com/lykke/matching/engine/database/azure/AzureMarketOrderDatabaseAccessor.kt
@@ -18,7 +18,11 @@ class AzureMarketOrderDatabaseAccessor  constructor(connectionString: String,
 
     override fun addLkkTrades(trades: List<LkkTrade>) {
         try {
-            batchInsertOrMerge(lkkTradesTable, trades.map { AzureLkkTrade(it.assetPair, it.clientId, it.price, it.volume, it.date) })
+            batchInsertOrMerge(lkkTradesTable, trades.map { AzureLkkTrade(it.assetPair,
+                    it.clientId,
+                    it.price.toDouble(),
+                    it.volume.toDouble(),
+                    it.date) })
         } catch(e: Exception) {
             LOGGER.error("Unable to add lkk trades, size: ${trades.size}", e)
             METRICS_LOGGER.logError( "Unable to add lkk trades, size: ${trades.size}", e)
