+++ src/test/kotlin/com/lykke/matching/engine/performance/MultiLimitOrderServicePerformanceTest.kt
@@ -10,6 +10,7 @@ import com.lykke.matching.engine.utils.MessageBuilder
 import com.lykke.matching.engine.utils.PrintUtils
 import org.junit.Ignore
 import org.junit.Test
+import java.math.BigDecimal
 import java.util.*
 
 @Ignore
@@ -51,19 +52,20 @@ class MultiLimitOrderServicePerformanceTest: AbstractPerformanceTest() {
 
         testBackOfficeDatabaseAccessor.addAsset(Asset("USD", 2))
         testBackOfficeDatabaseAccessor.addAsset(Asset("EUR", 2))
-        testDictionariesDatabaseAccessor.addAssetPair(AssetPair("EURUSD", "EUR", "USD", 5, 0.1, 0.2))
+        testDictionariesDatabaseAccessor.addAssetPair(AssetPair("EURUSD", "EUR", "USD", 5,
+                BigDecimal.valueOf(0.1), BigDecimal.valueOf(0.2)))
 
         initServices()
 
         counter.executeAction { multiLimitOrderService.processMessage(MessageBuilder.buildMultiLimitOrderWrapper(pair = "EURUSD", clientId = "Client1",
                 volumes = listOf(
-                        VolumePrice(0.1, 2.0),
-                        VolumePrice(0.1, 1.5),
-                        VolumePrice(0.09, 1.3),
-                        VolumePrice(1.0, 1.2),
-                        VolumePrice(-1.0, 2.1),
-                        VolumePrice(-0.09, 2.2),
-                        VolumePrice(-0.1, 2.4)
+                        VolumePrice(BigDecimal.valueOf(0.1), BigDecimal.valueOf(2.0)),
+                        VolumePrice(BigDecimal.valueOf(0.1), BigDecimal.valueOf(1.5)),
+                        VolumePrice(BigDecimal.valueOf(0.09),BigDecimal.valueOf( 1.3)),
+                        VolumePrice(BigDecimal.valueOf(1.0), BigDecimal.valueOf(1.2)),
+                        VolumePrice(BigDecimal.valueOf(-1.0),BigDecimal.valueOf(2.1)),
+                        VolumePrice(BigDecimal.valueOf(-0.09), BigDecimal.valueOf(2.2)),
+                        VolumePrice(BigDecimal.valueOf(-0.1), BigDecimal.valueOf(2.4))
                 ),
                 ordersFee = listOf(),
                 ordersFees = listOf()))
@@ -78,8 +80,8 @@ class MultiLimitOrderServicePerformanceTest: AbstractPerformanceTest() {
         counter.executeAction {  multiLimitOrderService.processMessage(
                 buildOldMultiLimitOrderWrapper(pair = "EURUSD",
                         clientId = "Client1",
-                        volumes = listOf(VolumePrice(100.0, 1.2),
-                                VolumePrice(100.0, 1.3))))}
+                        volumes = listOf(VolumePrice(BigDecimal.valueOf(100.0), BigDecimal.valueOf(1.2)),
+                                VolumePrice(BigDecimal.valueOf(100.0), BigDecimal.valueOf(1.3)))))}
         return counter.getAverageTime()
     }
 
@@ -88,9 +90,11 @@ class MultiLimitOrderServicePerformanceTest: AbstractPerformanceTest() {
         initServices()
 
         counter.executeAction {  multiLimitOrderService
-                .processMessage(buildOldMultiLimitOrderWrapper(pair = "EURUSD", clientId = "Client1", volumes = listOf(VolumePrice(100.0, 1.2), VolumePrice(100.0, 1.3))))}
+                .processMessage(buildOldMultiLimitOrderWrapper(pair = "EURUSD", clientId = "Client1", volumes = listOf(VolumePrice(BigDecimal.valueOf(100.0), BigDecimal.valueOf(1.2)),
+                        VolumePrice(BigDecimal.valueOf(100.0), BigDecimal.valueOf(1.3)))))}
         counter.executeAction { multiLimitOrderService
-                .processMessage(buildOldMultiLimitOrderWrapper(pair = "EURUSD", clientId = "Client1", volumes = listOf(VolumePrice(100.0, 1.4), VolumePrice(100.0, 1.5)))) }
+                .processMessage(buildOldMultiLimitOrderWrapper(pair = "EURUSD", clientId = "Client1", volumes = listOf(VolumePrice(BigDecimal.valueOf(100.0), BigDecimal.valueOf(1.4)),
+                        VolumePrice(BigDecimal.valueOf(100.0), BigDecimal.valueOf(1.5))))) }
 
         return counter.getAverageTime()
     }
@@ -100,11 +104,14 @@ class MultiLimitOrderServicePerformanceTest: AbstractPerformanceTest() {
 
         initServices()
         counter.executeAction {  multiLimitOrderService
-                .processMessage(buildOldMultiLimitOrderWrapper(pair = "EURUSD", clientId = "Client1", volumes = listOf(VolumePrice(100.0, 1.2), VolumePrice(100.0, 1.3))))}
+                .processMessage(buildOldMultiLimitOrderWrapper(pair = "EURUSD", clientId = "Client1", volumes =
+                listOf(VolumePrice(BigDecimal.valueOf(100.0), BigDecimal.valueOf(1.2)), VolumePrice(BigDecimal.valueOf(100.0), BigDecimal.valueOf(1.3)))))}
         counter.executeAction {  multiLimitOrderService
-                .processMessage(buildOldMultiLimitOrderWrapper(pair = "EURUSD", clientId = "Client1", volumes = listOf(VolumePrice(100.0, 1.4), VolumePrice(100.0, 1.5))))}
+                .processMessage(buildOldMultiLimitOrderWrapper(pair = "EURUSD", clientId = "Client1", volumes =
+                listOf(VolumePrice(BigDecimal.valueOf(100.0), BigDecimal.valueOf(1.4)), VolumePrice(BigDecimal.valueOf(100.0), BigDecimal.valueOf(1.5)))))}
         counter.executeAction {  multiLimitOrderService
-                .processMessage(buildOldMultiLimitOrderWrapper(pair = "EURUSD", clientId = "Client1", volumes = listOf(VolumePrice(100.0, 2.0), VolumePrice(100.0, 2.1)), cancel = true))}
+                .processMessage(buildOldMultiLimitOrderWrapper(pair = "EURUSD", clientId = "Client1", volumes =
+                listOf(VolumePrice(BigDecimal.valueOf(100.0), BigDecimal.valueOf(2.0)), VolumePrice(BigDecimal.valueOf(100.0), BigDecimal.valueOf(2.1))), cancel = true))}
 
         return counter.getAverageTime()
     }
@@ -113,15 +120,15 @@ class MultiLimitOrderServicePerformanceTest: AbstractPerformanceTest() {
         val counter = ActionTimeCounter()
 
         initServices()
-        counter.executeAction { multiLimitOrderService.processMessage(buildOldMultiLimitOrderWrapper(pair = "EURUSD", clientId = "Client1", volumes = listOf(VolumePrice(100.0, 1.3),
-                VolumePrice(100.0, 1.2)))) }
+        counter.executeAction { multiLimitOrderService.processMessage(buildOldMultiLimitOrderWrapper(pair = "EURUSD", clientId = "Client1", volumes = listOf(VolumePrice(BigDecimal.valueOf(100.0), BigDecimal.valueOf(1.3)),
+                VolumePrice(BigDecimal.valueOf(100.0), BigDecimal.valueOf(1.2))))) }
 
         singleLimitOrderService.processMessage(MessageBuilder.buildLimitOrderWrapper(MessageBuilder.buildLimitOrder(clientId = "Client2", price = 1.25, volume = -150.0)))
 
         counter.executeAction { multiLimitOrderService.processMessage(buildOldMultiLimitOrderWrapper(pair = "EURUSD", clientId = "Client1",
-                volumes = listOf(VolumePrice(10.0, 1.3),
-                        VolumePrice(100.0, 1.26),
-                        VolumePrice(100.0, 1.2)), cancel = true)) }
+                volumes = listOf(VolumePrice(BigDecimal.valueOf(10.0), BigDecimal.valueOf(1.3)),
+                        VolumePrice(BigDecimal.valueOf(100.0), BigDecimal.valueOf(1.26)),
+                        VolumePrice(BigDecimal.valueOf(100.0), BigDecimal.valueOf(1.2))), cancel = true)) }
 
         return counter.getAverageTime()
     }
@@ -130,17 +137,17 @@ class MultiLimitOrderServicePerformanceTest: AbstractPerformanceTest() {
         val counter = ActionTimeCounter()
 
         initServices()
-        counter.executeAction { multiLimitOrderService.processMessage(buildOldMultiLimitOrderWrapper(pair = "EURUSD", clientId = "Client1", volumes = listOf(VolumePrice(-100.0, 1.2),
-                VolumePrice(-100.0, 1.3)))) }
+        counter.executeAction { multiLimitOrderService.processMessage(buildOldMultiLimitOrderWrapper(pair = "EURUSD", clientId = "Client1", volumes = listOf(VolumePrice(BigDecimal.valueOf(-100.0), BigDecimal.valueOf(1.2)),
+                VolumePrice(BigDecimal.valueOf(-100.0), BigDecimal.valueOf(1.3))))) }
 
 
         singleLimitOrderService.processMessage(MessageBuilder.buildLimitOrderWrapper(MessageBuilder.buildLimitOrder(clientId = "Client2", price = 1.25, volume = 150.0)))
 
         counter.executeAction { multiLimitOrderService.processMessage(buildOldMultiLimitOrderWrapper(pair = "EURUSD", clientId = "Client1",
-                volumes = listOf(VolumePrice(-10.0, 1.2),
-                        VolumePrice(-10.0, 1.24),
-                        VolumePrice(-10.0, 1.29),
-                        VolumePrice(-10.0, 1.3)), cancel = true)) }
+                volumes = listOf(VolumePrice(BigDecimal.valueOf(-10.0), BigDecimal.valueOf(1.2)),
+                        VolumePrice(BigDecimal.valueOf(-10.0), BigDecimal.valueOf(1.24)),
+                        VolumePrice(BigDecimal.valueOf(-10.0), BigDecimal.valueOf(1.29)),
+                        VolumePrice(BigDecimal.valueOf(-10.0), BigDecimal.valueOf(1.3))), cancel = true)) }
 
 
         return counter.getAverageTime()
@@ -155,12 +162,14 @@ class MultiLimitOrderServicePerformanceTest: AbstractPerformanceTest() {
 
         initServices()
 
-        counter.executeAction { multiLimitOrderService.processMessage(buildOldMultiLimitOrderWrapper(pair = "TIMEUSD", clientId = "Client5", volumes = listOf(VolumePrice(-100.0, 26.955076)))) }
-        counter.executeAction {multiLimitOrderService.processMessage(buildOldMultiLimitOrderWrapper(pair = "TIMEUSD", clientId = "Client5", volumes = listOf(VolumePrice(0.69031943, 26.915076))))}
+        counter.executeAction { multiLimitOrderService.processMessage(buildOldMultiLimitOrderWrapper(pair = "TIMEUSD", clientId = "Client5", volumes =
+        listOf(VolumePrice(BigDecimal.valueOf(-100.0), BigDecimal.valueOf(26.955076))))) }
+        counter.executeAction {multiLimitOrderService.processMessage(buildOldMultiLimitOrderWrapper(pair = "TIMEUSD", clientId = "Client5", volumes =
+        listOf(VolumePrice(BigDecimal.valueOf(0.69031943), BigDecimal.valueOf(26.915076)))))}
 
         singleLimitOrderService.processMessage(MessageBuilder.buildLimitOrderWrapper(MessageBuilder.buildLimitOrder(assetId = "TIMEUSD", clientId = "Client2", price = 26.88023, volume = -26.0)))
         counter.executeAction {multiLimitOrderService.processMessage(buildOldMultiLimitOrderWrapper(pair = "TIMEUSD", clientId = "Client5",
-                volumes = listOf(VolumePrice(10.0, 26.915076), VolumePrice(10.0, 26.875076)), cancel = true))}
+                volumes = listOf(VolumePrice(BigDecimal.valueOf(10.0), BigDecimal.valueOf(26.915076)), VolumePrice(BigDecimal.valueOf(10.0), BigDecimal.valueOf(26.875076))), cancel = true))}
 
         return counter.getAverageTime()
     }
@@ -175,18 +184,23 @@ class MultiLimitOrderServicePerformanceTest: AbstractPerformanceTest() {
 
         singleLimitOrderService.processMessage(MessageBuilder.buildLimitOrderWrapper(MessageBuilder.buildLimitOrder(assetId = "BTCEUR", clientId = "Client2", price = 3629.355, volume = 0.19259621)))
 
-        counter.executeAction { multiLimitOrderService.processMessage(buildOldMultiLimitOrderWrapper(pair = "BTCEUR", clientId = "Client5", volumes = listOf(VolumePrice(-0.00574996, 3628.707)), cancel = true)) }
+        counter.executeAction { multiLimitOrderService.processMessage(buildOldMultiLimitOrderWrapper(pair = "BTCEUR", clientId = "Client5", volumes =
+        listOf(VolumePrice(BigDecimal.valueOf(-0.00574996), BigDecimal.valueOf(3628.707))), cancel = true)) }
         counter.executeAction { multiLimitOrderService.processMessage(buildOldMultiLimitOrderWrapper(pair = "BTCEUR", clientId = "Client5",
-                volumes = listOf(VolumePrice(-0.01431186, 3624.794), VolumePrice(-0.02956591, 3626.591)), cancel = true))}
-        counter.executeAction {multiLimitOrderService.processMessage(buildOldMultiLimitOrderWrapper(pair = "BTCEUR", clientId = "Client5", volumes = listOf(VolumePrice(-0.04996673, 3625.855)), cancel = true))}
+                volumes = listOf(VolumePrice(BigDecimal.valueOf(-0.01431186), BigDecimal.valueOf(3624.794)),
+                        VolumePrice(BigDecimal.valueOf(-0.02956591), BigDecimal.valueOf(3626.591))), cancel = true))}
+        counter.executeAction {multiLimitOrderService.processMessage(buildOldMultiLimitOrderWrapper(pair = "BTCEUR", clientId = "Client5", volumes =
+        listOf(VolumePrice(BigDecimal.valueOf(-0.04996673), BigDecimal.valueOf(3625.855))), cancel = true))}
         counter.executeAction {multiLimitOrderService.processMessage(buildOldMultiLimitOrderWrapper(pair = "BTCEUR", clientId = "Client5",
-                volumes = listOf(VolumePrice(-0.00628173, 3622.865),
-                        VolumePrice(-0.01280207, 3625.489),
-                        VolumePrice(-0.02201331, 3627.41),
-                        VolumePrice(-0.02628901, 3629.139)), cancel = true))}
+                volumes = listOf(VolumePrice(BigDecimal.valueOf(-0.00628173), BigDecimal.valueOf(3622.865)),
+                        VolumePrice(BigDecimal.valueOf(-0.01280207), BigDecimal.valueOf(3625.489)),
+                        VolumePrice(BigDecimal.valueOf(-0.02201331), BigDecimal.valueOf(3627.41)),
+                        VolumePrice(BigDecimal.valueOf(-0.02628901), BigDecimal.valueOf(3629.139))), cancel = true))}
 
-        counter.executeAction {multiLimitOrderService.processMessage(buildOldMultiLimitOrderWrapper(pair = "BTCEUR", clientId = "Client5", volumes = listOf(VolumePrice(-0.01708411, 3626.11)), cancel = true))}
-        counter.executeAction {multiLimitOrderService.processMessage(buildOldMultiLimitOrderWrapper(pair = "BTCEUR", clientId = "Client5", volumes = listOf(VolumePrice(-0.00959341, 3625.302)), cancel = true))}
+        counter.executeAction {multiLimitOrderService.processMessage(buildOldMultiLimitOrderWrapper(pair = "BTCEUR", clientId = "Client5", volumes =
+        listOf(VolumePrice(BigDecimal.valueOf(-0.01708411), BigDecimal.valueOf(3626.11))), cancel = true))}
+        counter.executeAction {multiLimitOrderService.processMessage(buildOldMultiLimitOrderWrapper(pair = "BTCEUR", clientId = "Client5", volumes =
+        listOf(VolumePrice(BigDecimal.valueOf(-0.00959341), BigDecimal.valueOf(3625.302))), cancel = true))}
 
         return counter.getAverageTime()
     }
@@ -205,10 +219,13 @@ class MultiLimitOrderServicePerformanceTest: AbstractPerformanceTest() {
         singleLimitOrderService.processMessage(MessageBuilder.buildLimitOrderWrapper(MessageBuilder.buildLimitOrder(clientId = "Client2", assetId = "BTCCHF", uid = "1", price = 4384.15, volume = -0.26070853)))
 
 
-        counter.executeAction { multiLimitOrderService.processMessage(buildOldMultiLimitOrderWrapper(pair = "BTCCHF", clientId = "Client3", volumes = listOf(VolumePrice(0.00643271, 4390.84),
-                VolumePrice(0.01359005, 4387.87), VolumePrice(0.02033985, 4384.811)), cancel = true)) }
+        counter.executeAction { multiLimitOrderService.processMessage(buildOldMultiLimitOrderWrapper(pair = "BTCCHF", clientId = "Client3", volumes =
+        listOf(VolumePrice(BigDecimal.valueOf(0.00643271), BigDecimal.valueOf(4390.84)),
+                VolumePrice(BigDecimal.valueOf(0.01359005), BigDecimal.valueOf(4387.87)),
+                VolumePrice(BigDecimal.valueOf(0.02033985), BigDecimal.valueOf(4384.811))), cancel = true)) }
 
-        counter.executeAction { multiLimitOrderService.processMessage(buildOldMultiLimitOrderWrapper(pair = "BTCCHF", clientId = "Client3", volumes = listOf(VolumePrice(0.01691068, 4387.21)), cancel = true))}
+        counter.executeAction { multiLimitOrderService.processMessage(buildOldMultiLimitOrderWrapper(pair = "BTCCHF", clientId = "Client3",
+                volumes = listOf(VolumePrice(BigDecimal.valueOf(0.01691068), BigDecimal.valueOf(4387.21))), cancel = true))}
 
         return counter.getAverageTime()
     }
@@ -226,7 +243,8 @@ class MultiLimitOrderServicePerformanceTest: AbstractPerformanceTest() {
         testOrderDatabaseAccessor.addLimitOrder(MessageBuilder.buildLimitOrder(clientId = client, assetId = "BTCEUR", price = 4722.0, volume = 0.14825226))
         initServices()
 
-        counter.executeAction { multiLimitOrderService.processMessage(buildOldMultiLimitOrderWrapper(pair = "BTCEUR", clientId = marketMaker, volumes = listOf(VolumePrice(-0.4435, 4721.403)), cancel = true)) }
+        counter.executeAction { multiLimitOrderService.processMessage(buildOldMultiLimitOrderWrapper(pair = "BTCEUR", clientId = marketMaker, volumes =
+        listOf(VolumePrice(BigDecimal.valueOf(-0.4435), BigDecimal.valueOf(4721.403))), cancel = true)) }
 
         return counter.getAverageTime()
     }
@@ -241,13 +259,13 @@ class MultiLimitOrderServicePerformanceTest: AbstractPerformanceTest() {
         testBalanceHolderWrapper.updateBalance(client, "USD", 1.19)
 
         val order = MessageBuilder.buildLimitOrder(clientId = client, assetId = "EURUSD", price = 1.2, volume = 1.0)
-        order.reservedLimitVolume = 1.19
+        order.reservedLimitVolume = BigDecimal.valueOf(1.19)
         testOrderDatabaseAccessor.addLimitOrder(order)
 
         initServices()
 
         counter.executeAction { multiLimitOrderService.processMessage(MessageBuilder.buildMultiLimitOrderWrapper(clientId = marketMaker, pair = "EURUSD",
-                volumes = listOf(VolumePrice(-2.0, 1.1)), cancel = false, ordersFee = listOf(), ordersFees = listOf())) }
+                volumes = listOf(VolumePrice(BigDecimal.valueOf(-2.0), BigDecimal.valueOf(1.1))), cancel = false, ordersFee = listOf(), ordersFees = listOf())) }
 
         return counter.getAverageTime()
     }
@@ -256,10 +274,10 @@ class MultiLimitOrderServicePerformanceTest: AbstractPerformanceTest() {
         val counter = ActionTimeCounter()
 
         counter.executeAction { multiLimitOrderService.processMessage(MessageBuilder.buildMultiLimitOrderWrapper(pair = "EURUSD", clientId = "Client1",
-                volumes = listOf(VolumePrice(-9.0, 0.4875)), ordersUid = listOf("order1"), cancel = true, ordersFee = emptyList(), ordersFees = emptyList())) }
+                volumes = listOf(VolumePrice(BigDecimal.valueOf(-9.0), BigDecimal.valueOf(0.4875))), ordersUid = listOf("order1"), cancel = true, ordersFee = emptyList(), ordersFees = emptyList())) }
 
         counter.executeAction { multiLimitOrderService.processMessage(MessageBuilder.buildMultiLimitOrderWrapper(pair = "EURUSD", clientId = "Client1",
-                volumes = listOf(VolumePrice(-10.0, 0.4880)), ordersUid = listOf("order1"), cancel = true, ordersFee = emptyList(), ordersFees = emptyList())) }
+                volumes = listOf(VolumePrice(BigDecimal.valueOf(-10.0), BigDecimal.valueOf(0.4880))), ordersUid = listOf("order1"), cancel = true, ordersFee = emptyList(), ordersFees = emptyList())) }
 
         return counter.getAverageTime()
     }
@@ -279,8 +297,8 @@ class MultiLimitOrderServicePerformanceTest: AbstractPerformanceTest() {
                 .setCancelAllPreviousLimitOrders(cancel)
         volumes.forEach{ volume ->
             orderBuilder.addOrders(ProtocolMessages.OldMultiLimitOrder.Order.newBuilder()
-                    .setVolume(volume.volume)
-                    .setPrice(volume.price)
+                    .setVolume(volume.volume.toDouble())
+                    .setPrice(volume.price.toDouble())
                     .build())
         }
         return orderBuilder.build()
